fix for long type cast

+++ src/common/slurmdb_defs.c
@@ -1638,7 +1638,7 @@ extern int slurmdb_report_set_start_end_time(time_t *start, time_t *end)
 	if(!sent_end) {
 		if(!localtime_r(&my_time, &end_tm)) {
 			error("Couldn't get localtime from end %ld",
-			      my_time);
+			      (long)my_time);
 			return SLURM_ERROR;
 		}
 		end_tm.tm_hour = 0;
@@ -1647,7 +1647,7 @@ extern int slurmdb_report_set_start_end_time(time_t *start, time_t *end)
 		temp_time = sent_end;
 		if(!localtime_r(&temp_time, &end_tm)) {
 			error("Couldn't get localtime from user end %ld",
-			      my_time);
+			      (long)my_time);
 			return SLURM_ERROR;
 		}
 		if(end_tm.tm_sec >= 30)
@@ -1664,7 +1664,7 @@ extern int slurmdb_report_set_start_end_time(time_t *start, time_t *end)
 	if(!sent_start) {
 		if(!localtime_r(&my_time, &start_tm)) {
 			error("Couldn't get localtime from start %ld",
-			      my_time);
+			      (long)my_time);
 			return SLURM_ERROR;
 		}
 		start_tm.tm_hour = 0;
@@ -1674,7 +1674,7 @@ extern int slurmdb_report_set_start_end_time(time_t *start, time_t *end)
 		temp_time = sent_start;
 		if(!localtime_r(&temp_time, &start_tm)) {
 			error("Couldn't get localtime from user start %ld",
-			      my_time);
+			      (long)my_time);
 			return SLURM_ERROR;
 		}
 		if(start_tm.tm_sec >= 30)
